--- app/controllers/forms/__init__.py
@@ -1,15 +1,0 @@
-# This is needed for ExtendedFlaskView to automatically import all Form classes
-
-from .ingredients import IngredientsForm  # noqa: F401
-from .events import EventsForm  # noqa: F401
-from .recipes import RecipesForm  # noqa: F401
-from .users import UsersForm  # noqa: F401
-from .public_recipes import PublicRecipeFilterForm  # noqa: F401
-
-__all__ = [
-    "IngredientsForm",
-    "EventsForm",
-    "RecipesForm",
-    "UsersForm",
-    "PublicRecipeFilterForm",
-]
