+++ app/forms/__init__.py
@@ -1,0 +1,16 @@
+from .events import EventsForm
+from .ingredients import IngredientsForm
+from .public_recipes import PublicRecipeFilterForm
+from .recipes import RecipesForm
+from .support import FeedbackForm
+from .users import UsersForm, SetPasswordForm
+
+__all__ = [
+    "IngredientsForm",
+    "EventsForm",
+    "RecipesForm",
+    "PublicRecipeFilterForm",
+    "FeedbackForm",
+    "UsersForm",
+    "SetPasswordForm",
+]
